Add files via upload
Everyone working with PDFs must have seen the CTRL+F option, where you can search
for a word and the output says how many such words are there in the PDF document. It
also offers the option of replacement.
So, write a function that takes a ".txt" file, a string, and another keyword argument
"operation" as input and does the following:
• If the operation is given as "find". Then it says the number of times that the input
string has been present in the ".txt file".
• If the operation is given as "replace," then it needs another keyword argument=,
"serial" and "replacement." If the serial is "i," then the i-th instance of the original
word is replaced by the "replacement" string. The serial must not be greater than
the number of instances of the original string that has been repeated in the ".txt"
file. Otherwise, the function should output an error.
• If the operation is given as "replace-all" then the keyword serial is not needed,
but the "replacement" keyword is needed. All the original instances of the word is
replaced by the replacement in the file.